// src/utils/interfaces/interfaces.ts
export interface IProductStatusPayload {
  isActive?: boolean;
  isAvailable?: boolean;
  isFeatured?: boolean;
  isNewProduct?: boolean;
  isBestSeller?: boolean;
  isLimitedEdition?: boolean;
  isExclusive?: boolean;
  isPreOrder?: boolean;
  isBackOrder?: boolean;
  isComingSoon?: boolean;
  isOutOfStock?: boolean;
  isDiscontinued?: boolean;
  isArchived?: boolean;
  isPromo?: boolean;
  isInAd?: boolean;
}